Move table rendering logic to SafeHtmlTable component
